Migrated to Vite:
- All CRA modules have been removed so just essential vite modules + eslint and prettier
- Eye SVGS don't work, need to figure out how to use svgs in vite
