--- .env.example
@@ -1,1 +1,0 @@
-REACT_APP_BACKEND_URL=http://localhost:5050
