--- jsconfig.json
@@ -1,7 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": "src",
-    "jsx": "react"
-  },
-  "include": ["src"]
-}
