+++ src/common/components/form/SubmitButton.jsx
@@ -1,4 +1,3 @@
-import React from 'react';
 
 import PropTypes from 'prop-types';
