+++ src/common/components/form/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { Button } from 'common/components/Button';
+import { Button } from '@/common/components/Button';
 
 export const InputContainer = styled.div``;
