+++ src/common/components/routes/ProtectedRoutes.jsx
@@ -1,8 +1,7 @@
-import React from 'react';
 
 import { Navigate, Outlet } from 'react-router-dom';
 
-import { useUser } from 'common/contexts/UserContext';
+import { useUser } from '@/common/contexts/UserContext';
 
 export function PrivateRoute() {
   const { user, isLoading } = useUser();
