+++ src/common/layouts/NavLayout.jsx
@@ -1,9 +1,8 @@
-import React from 'react';
 
 import { Outlet } from 'react-router-dom';
 import styled from 'styled-components';
 
-import NavBar from 'common/components/navigation/NavBar';
+import NavBar from '@/common/components/navigation/NavBar';
 
 const Layout = styled.div`
   height: 100vh;
