+++ src/pages/account/ResetPassword.jsx
@@ -1,9 +1,9 @@
-import React, { useEffect, useState } from 'react';
+import { useEffect, useState } from 'react';
 
 import { useNavigate } from 'react-router-dom';
 import styled from 'styled-components';
 
-import { useUser } from 'common/contexts/UserContext';
+import { useUser } from '@/common/contexts/UserContext';
 
 const Container = styled.div`
   max-width: 400px;
