+++ src/pages/account/SignUp.jsx
@@ -1,12 +1,12 @@
-import React, { useState } from 'react';
+import { useState } from 'react';
 
 import { useNavigate } from 'react-router-dom';
 
-import GoogleButton from 'common/components/GoogleButton';
-import { Form, FormTitle } from 'common/components/form/Form';
-import { Input } from 'common/components/form/Input';
-import SubmitButton from 'common/components/form/SubmitButton';
-import { useUser } from 'common/contexts/UserContext';
+import GoogleButton from '@/common/components/GoogleButton';
+import { Form, FormTitle } from '@/common/components/form/Form';
+import { Input } from '@/common/components/form/Input';
+import SubmitButton from '@/common/components/form/SubmitButton';
+import { useUser } from '@/common/contexts/UserContext';
 
 import { StyledPage } from './styles';
