+++ src/pages/account/styles.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import { Button } from 'common/components/Button';
+import { Button } from '@/common/components/Button';
 
 export const StyledForm = styled.form`
   display: flex;
