+++ vite.config.ts
@@ -1,0 +1,16 @@
+import { defineConfig } from 'vite'
+import react from '@vitejs/plugin-react'
+import path from 'path';
+
+// https://vite.dev/config/
+export default defineConfig({
+  plugins: [react()],
+  server: {
+    port: 3001,
+  },
+  resolve: {
+    alias: {
+      '@': path.resolve(__dirname, "src"),
+    }
+  }
+})
